Rearrange dir structure for convenience classes
Like the middleware. The idea is that all wrappers/middleware/etc that
make life easier for the user of the library will share a namespace.
Looking ahead at the wrappers, the OpenTransaction looks like a class
that will be shared amongst those convenience classes so we moved it up
a level.