Buffer calls in HttplugAsyncClient and send async
Any caller of the async client was still waiting for responses, making
the async nature a bit of a farce. Here, wee move the waiting for
responses to a function that is called on shutdown, after the whole
script has executed. This makes it possible for all callers to just
send transactions and errors and leave the waiting for responses to the
client, which triggers it at the last possible moment. Nice side effect
is that we can now implement the simpler CLient interface since the
async nature is abstracted away.